Lineas duplicadas eliminadas

+++ api/sistemas_transporte/delete.php
@@ -25,10 +25,15 @@ $data = json_decode(file_get_contents("php://input"));
 $transport->id = $data->id;
  
 // delete the product
-if($transport->delete()){
-    echo json_encode(array("message" => "Product was deleted."));
-}else{
-    echo json_encode(array("message" => "Unable to delete product."));
+if($data->id != null){
+    if($transport->delete()){
+        echo json_encode(array("message" => "Product was deleted."));
+    }else{
+        echo json_encode(array("message" => "Unable to delete product."));
+    }
+} else {
+    echo json_encode(array("message" => "Faltan datos"));
 }
+
 
 ?>
